requirements.txt
opencv-python
mediapipe
python-chess
numpy